# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "오뚜기 파워세일즈 API",
    "description": "B2B 영업사원 실적 조회 및 목표 관리 시스템 API",
    "version": "v1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:49275",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "Bearer Authentication": []
    }
  ],
  "paths": {
...
    "/api/v1/mobile/mypage/schedule/monthly": {
      "get": {
        "tags": [
          "my-schedule-controller"
        ],
        "operationId": "getMonthlySchedule",
        "parameters": [
          {
            "name": "year",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "month",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseMonthlyScheduleResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/mobile/mypage/schedule/daily": {
      "get": {
        "tags": [
          "my-schedule-controller"
        ],
        "operationId": "getDailySchedule",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDailyScheduleResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/team-schedule/accounts": {
      "get": {
        "tags": [
          "admin-team-schedule-controller"
        ],
        "operationId": "getAccounts",
        "parameters": [
          {
            "name": "branchCode",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListTeamScheduleAccountDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/accounts": {
      "get": {
        "tags": [
          "admin-account-controller"
        ],
        "operationId": "getAccounts_1",
        "parameters": [
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 50,
              "minLength": 1
            }
          },
          {
            "name": "abcType",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "branchCode",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "accountStatusName",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0,
              "minimum": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20,
              "maximum": 100,
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseAccountListResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorDetail": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
...
      "ApiResponseMonthlyScheduleResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/MonthlyScheduleResponse"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "message": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "success",
          "timestamp"
        ]
      },
      "MonthlyScheduleResponse": {
        "type": "object",
        "properties": {
          "year": {
            "type": "integer",
            "format": "int32"
          },
          "month": {
            "type": "integer",
            "format": "int32"
          },
          "work_days": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkDayDto"
            }
          },
          "annualLeaveCount": {
            "type": "integer",
            "format": "int32"
          }
        },
        "required": [
          "annualLeaveCount",
          "month",
          "work_days",
          "year"
        ]
      },
      "WorkDayDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string"
          },
          "has_work": {
            "type": "boolean"
          },
          "workingType": {
            "type": "string"
          }
        },
        "required": [
          "date",
          "has_work"
        ]
      },
      "ApiResponseDailyScheduleResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/DailyScheduleResponse"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "message": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "success",
          "timestamp"
        ]
      },
      "DailyScheduleResponse": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string"
          },
          "day_of_week": {
            "type": "string"
          },
          "member_name": {
            "type": "string"
          },
          "employee_code": {
            "type": "string"
          },
          "report_progress": {
            "$ref": "#/components/schemas/ReportProgressDto"
          },
          "accounts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DisplayWorkScheduleItemDto"
            }
          }
        },
        "required": [
          "accounts",
          "date",
          "day_of_week",
          "employee_code",
          "member_name",
          "report_progress"
        ]
      },
      "DisplayWorkScheduleItemDto": {
        "type": "object",
        "properties": {
          "account_id": {
            "type": "integer",
            "format": "int64"
          },
          "account_name": {
            "type": "string"
          },
          "work_type_1": {
            "type": "string"
          },
          "work_type_2": {
            "type": "string"
          },
          "work_type_3": {
            "type": "string"
          },
          "is_registered": {
            "type": "boolean"
          }
        },
        "required": [
          "account_id",
          "account_name",
          "is_registered",
          "work_type_1",
          "work_type_2",
          "work_type_3"
        ]
      },
      "ReportProgressDto": {
        "type": "object",
        "properties": {
          "completed": {
            "type": "integer",
            "format": "int32"
          },
          "total": {
            "type": "integer",
            "format": "int32"
          },
          "work_type": {
            "type": "string"
          }
        },
        "required": [
          "completed",
          "total",
          "work_type"
        ]
      },
...
      "AccountInfo": {
        "type": "object",
        "properties": {
          "scheduleId": {
            "type": "integer",
            "format": "int64"
          },
          "accountSfid": {
            "type": "string"
          },
          "accountName": {
            "type": "string"
          },
          "accountTypeCode": {
            "type": "string"
          },
          "workCategory": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "latitude": {
            "type": "number",
            "format": "double"
          },
          "longitude": {
            "type": "number",
            "format": "double"
          },
          "isRegistered": {
            "type": "boolean"
          }
        },
        "required": [
          "accountName",
          "isRegistered",
          "scheduleId",
          "workCategory"
        ]
      },
      "AccountListResponse": {
        "type": "object",
        "properties": {
          "accounts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AccountInfo"
            }
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "registeredCount": {
            "type": "integer",
            "format": "int32"
          },
          "currentDate": {
            "type": "string"
          }
        },
        "required": [
          "accounts",
          "currentDate",
          "registeredCount",
          "totalCount"
        ]
      },
      "ApiResponseAccountListResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/AccountListResponse"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "message": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "success",
          "timestamp"
        ]
      },
...
      "ApiResponseListTeamScheduleAccountDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TeamScheduleAccountDto"
            }
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "message": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "success",
          "timestamp"
        ]
      },
      "TeamScheduleAccountDto": {
        "type": "object",
        "properties": {
          "accountId": {
            "type": "integer",
            "format": "int32"
          },
          "externalKey": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "accountId",
          "externalKey",
          "name"
        ]
      },
...
    },
    "securitySchemes": {
      "Bearer Authentication": {
        "type": "http",
        "name": "Bearer Authentication",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}